suddivision di build in build, buld_arch, build_debian e build_fedora
