+++ coa/docs/md/coa.md
@@ -23,4 +23,4 @@ coa is the orchestrator written in Go, designed to be the "Mind" behind the GNU/
 * [coa sysinstall](coa_sysinstall.md)	 - Install the live system to a physical disk
 * [coa version](coa_version.md)	 - Print the version number of coa
 
-###### Auto generated by spf13/cobra on 14-Apr-2026
+###### Auto generated by spf13/cobra on 15-Apr-2026
