+++ coa/docs/md/coa_build.md
@@ -28,4 +28,4 @@ coa build [flags]
 
 * [coa](coa.md)	 - coa (brooding in my dialect) - The Mind orchestrator
 
-###### Auto generated by spf13/cobra on 14-Apr-2026
+###### Auto generated by spf13/cobra on 15-Apr-2026
