+++ coa/docs/md/coa_export.md
@@ -19,4 +19,4 @@ Export generated ISOs or native packages to a remote server via SCP.
 * [coa export iso](coa_export_iso.md)	 - Export the latest ISO to a remote Proxmox storage
 * [coa export pkg](coa_export_pkg.md)	 - Export the latest generated native package (.deb or .pkg.tar.zst) to Proxmox
 
-###### Auto generated by spf13/cobra on 14-Apr-2026
+###### Auto generated by spf13/cobra on 15-Apr-2026
