+++ coa/docs/md/coa_export_pkg.md
@@ -22,4 +22,4 @@ coa export pkg [flags]
 
 * [coa export](coa_export.md)	 - Export artifacts (iso, pkg) to a remote Proxmox storage
 
-###### Auto generated by spf13/cobra on 14-Apr-2026
+###### Auto generated by spf13/cobra on 15-Apr-2026
