+++ coa/docs/md/coa_sysinstall.md
@@ -30,4 +30,4 @@ coa sysinstall [flags]
 
 * [coa](coa.md)	 - coa (brooding in my dialect) - The Mind orchestrator
 
-###### Auto generated by spf13/cobra on 14-Apr-2026
+###### Auto generated by spf13/cobra on 15-Apr-2026
