+++ coa/docs/md/coa_version.md
@@ -16,4 +16,4 @@ coa version [flags]
 
 * [coa](coa.md)	 - coa (brooding in my dialect) - The Mind orchestrator
 
-###### Auto generated by spf13/cobra on 14-Apr-2026
+###### Auto generated by spf13/cobra on 15-Apr-2026
